add hlsl backend evaluator for WSL 1.0

+++ engine/install_antares_host.sh
@@ -1,10 +1,23 @@
-#!/bin/sh -ex
+#!/bin/bash -e
 
 cd $(dirname $0)/..
 ANTARES_ROOT=$(pwd)
 
-dpkg -L git python3-dev python3-pip g++ llvm-dev make >/dev/null 2>&1 || \
-  apt-get update && apt-get install -y --no-install-recommends git python3-dev python3-pip g++ llvm-dev make
+VERSION_TAG=v0.2dev0
+
+REQUIRED_PACKAGES="git python3-dev python3-pip g++ llvm-dev make curl libopenmpi-dev openmpi-bin"
+
+if grep Microsoft /proc/sys/kernel/osrelease >/dev/null; then
+  REQUIRED_PACKAGES="${REQUIRED_PACKAGES} g++-mingw-w64-x86-64"
+fi
+
+if [[ "$(whoami)" != "root" ]]; then
+  echo "Root previledge is required for dependency installation."
+  exit 1
+fi
+
+dpkg -L ${REQUIRED_PACKAGES} >/dev/null 2>&1 || \
+  sh -c "apt-get update && apt-get install -y --no-install-recommends ${REQUIRED_PACKAGES}"
 
 TVM_HOME=/opt/tvm
 rm -rf $TVM_HOME && git clone https://github.com/apache/incubator-tvm $TVM_HOME
@@ -21,3 +34,4 @@ cd $TVM_HOME && git checkout 73f425d && git apply device-stub/tvm_v0.7.patch && 
 
 pip3 install --upgrade tornado psutil xgboost==1.2.1 numpy decorator attrs pytest typed_ast mpi4py
 
+echo "$VERSION_TAG" > $TVM_HOME/VERSION_TAG
